Changes to allow GLD working
Made two changes:

- Adding line to allow `harmonization` to take value `"GLD"` (new line `453`)
- Dropped `wgt` from local `I2D2_varlist`

The first is because it was missing. The second is because line `471` does `rename weight wgt`. If `wgt` is part of local `I2D2_varlist` then `wgt` already exists (either alreday in I2D2 or create in the loop following the local definition). If it exists then we cannot rename (Stata stops with `wgt already defined`)